requirements.txt
asgiref==3.9.1
Django==4.2.23
sqlparse==0.5.3
wheel==0.45.1

# Database
psycopg2-binary==2.9.9

# API Framework
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-filter==25.1

# Task Queue
celery==5.3.4
redis==5.0.1

# File Processing
pandas>=2.0.0
openpyxl==3.1.2
xlrd==2.0.1
chardet==5.2.0

# Deep Learning Framework (PyTorch-based)
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Transformers for NLP
transformers>=4.30.0
tokenizers>=0.13.0

# Machine Learning
scikit-learn>=1.3.0
numpy>=1.24.0
scipy>=1.10.0

# Data Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Optional: CUDA support (uncomment if you have NVIDIA GPU)
# torch-audio>=2.0.0+cu118
# torch-vision>=0.15.0+cu118
# torch>=2.0.0+cu118

# Documentation
drf-spectacular==0.27.0

# Utilities
python-decouple==3.8
django-cors-headers==4.3.1
django-extensions==3.2.3

# Reports
reportlab==4.0.7
WeasyPrint==60.2

# Development
django-debug-toolbar==4.2.0
factory-boy==3.3.0
